carnot-attest/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "carnot-attest"
version = "0.1.1"
description = "Generate signed attestations from CryptoBOM v2.1"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name="Carnot Engine"}]
dependencies = ["pynacl>=1.5.0","jinja2>=3.1"]

[project.scripts]
carnot-attest = "carnot_attest.__main__:main"

[tool.setuptools]
packages = ["carnot_attest"]
